Enhance TryGetLastActor to support HTTP request input (#635)
* Enhance TryGetLastActor to support HTTP request input

Updated the `TryGetLastActor` method in `ActivityExecutionContextExtensions.cs` to retrieve the last actor from the HTTP request body if an "actor" field is present. If not, it falls back to the Workflow context variable "RunAsSystemUser" and the last actor from the Variables. Adjusted using directives and updated comments to reflect these changes.

* Updated changelog for identity

* Improve actor retrieval and update changelog for 8.0.0-rc26

- Enhanced `ActivityExecutionContextExtensions.cs` to prioritize system user retrieval with a new `forceRunAsSystemUser` variable.
- Removed the previous `runAsSystemUser` check for streamlined logic.
- Updated `CHANGELOG.md` for version `8.0.0-rc26`, adding support for configurable per-user Two Factor enforcement policy.
- Removed session state dependency; recommend removing `app.UseSession()` unless necessary.
- Added migration for a new column in the user table.

* Add TrySetLastActor method and improve documentation

Implemented TrySetLastActor in HttpEndpointWithValidation.cs to set the last actor from the HTTP request body. Updated ExecuteInternal to utilize this new method. Enhanced TryGetLastActor in ActivityExecutionContextExtensions.cs for better logic and documentation clarity.