fixes in GetServerLoopbackUri & case access rule predicate (#667)
* Update GetServerLoopbackUri to handle server not running at startup

Modified the `GetServerLoopbackUri` method in the `HttpClientFactoryExtensions` class to return a nullable `Uri`. The method now retrieves the server's address and falls back to the `ASPNETCORE_URLS` configuration if no address is found. This change improves robustness by allowing the method to return `null` when no valid address is available, preventing potential exceptions.

* Refactor AccessMatchPredicate userId comparison

Updated the `AccessMatchPredicate` method in the `DbCaseAccessRule` class to simplify the userId comparison by removing the unnecessary `ToString()` conversion. The comparison now directly uses the nullable string type for improved clarity and efficiency.

* Code QL fixes

* Add conditional certificate loading for .NET 9.0+

Introduces a compilation directive to use `X509CertificateLoader.LoadCertificate` for .NET 9.0 or greater. Falls back to `X509Certificate2` for earlier versions, ensuring compatibility with new features while maintaining support for older frameworks.

* nullables

* Enhance sign-in tracking

- Added `LastSignInDate` property to `ExtendedSignInManager` for better device tracking.
- Updated `ExtendedResourceOwnerPasswordValidator` to set last sign-in date on successful login.
- Modified `ExtendedTokenResponseGenerator` to handle deprecated types for .NET compatibility.

* Fix URI construction in HttpClientFactoryExtensions

Added System.Net namespace and updated URI logic to replace
"+" with "0.0.0.0" in addresses, ensuring proper formatting
for URI parsing.

* remove namespace

* Improve error handling for username change failures

Updated the assertion for username changes to include detailed error information. Instead of a generic failure message, the new implementation captures the first error from the result, providing both the error code and description in the assertion failure message.

* Fix test to teload TestUser after device registration

Added code to retrieve the `ExtendedIdentityDbContext` and reload the `TestUser` entity from the database. This ensures that any changes made to the user in the database are reflected in the test context before proceeding to change the username.

* Refactor validation for NewPasswordConfirmation

Separated validation rules for NewPasswordConfirmation to ensure it is not empty and equals NewPassword. Added localized name for better clarity in validation messages.

---------

Co-authored-by: Constantinos Leftheris <c.leftheris@live.com>