Add IHostEnvironment support to SmsServiceSmsUp
Integrated IHostEnvironment to include hosting environment
details in the SmsServiceSmsUp class. Added the `Environment`
property and updated the constructor to accept an
IHostEnvironment parameter. Modified the `Custom` property
in SMS messages to include the application name and
environment name for better tracking and debugging.
Added necessary namespaces to support new dependencies.