[cases]  private data draft (#858)
* Refactor public data selection logic in case services

Introduce a local publicData variable in BaseCaseService to clarify and centralize the logic for selecting public vs. internal case data, ensuring public data is only fetched for non-draft cases. Update LINQ projections and adjust MyCaseService to use the revised method signature for improved readability and maintainability.

* Update version and allow custom ServerLoopbackUri config

Increment VersionPrefixIdentity and VersionPrefixIdentityUI to .1.
Enhance HttpClient setup in CaseServerFeatureExtensions and CasesWorkflowFeatureExtensions to use a configurable "ServerLoopbackUri" if provided, falling back to the default loopback URI otherwise.